backend/requirements.txt
django==4.1.7
python-dotenv==1.0.0
django-colorfield==0.8.0
django-nested-admin==4.0.2
pytest==7.2.2
pytest-django==4.5.2
gunicorn==20.0.4
psycopg2